requirements.txt
schedule
SRTrain
discord
python-dotenv